WIP Initial Docker devstack configuration
This commit adds a basic Docker devstack configuration for the Django
Boston website. The goal is to create a very minimal site using
production infrastructure to get the CI/CD work done first

As of this commit, we can use `docker-compose` to run a django app with
Nginx, Gunicorn, and Postgresql.

TO FIX: Select the database depending on if we are running with a
Postgresql database (as is done in Docker) or with SQLite, which should
be run for unit tests and the default for standalone app mode (not
running with Postgresql)